Update logsync.sh
added "--ignore-checksum" flag because the updated log file would never match the old. For some reason the old script worked for about 2 days without matching the MD5 checksums but once the log grew to large it generated errors. Unfortunately, the ignore checksum flag is not available on release 1.35 so the version in the apt-get repository does not work. I had to install from  https://downloads.rclone.org/rclone-current-linux-arm.zip and re-configure rclone.